fixed zonal_stats None value return; added option for ssl_verify=False
zonal_stats returns None if no centroids are within a polygon and all_touched=False
set all_touched=True as default. want to add in option to specify this eventually

jpl snow server request was having an issue with ssl certs. set verify=False and option to set this as intermediate fix.